Staff list is now a separate component in the components folder.
